requirements.txt
streamlit>=1.28.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
